+++ core/src/main/java/org/eclipse/hono/util/AdapterInstanceStatus.java
@@ -1,0 +1,36 @@
+/*******************************************************************************
+ * Copyright (c) 2021 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+
+package org.eclipse.hono.util;
+
+/**
+ * The adapter instance status.
+ */
+public enum AdapterInstanceStatus {
+    /**
+     * Adapter instance is alive.
+     */
+    ALIVE,
+    /**
+     * Adapter instance has probably been terminated.
+     */
+    SUSPECTED_DEAD,
+    /**
+     * Adapter instance has been terminated.
+     */
+    DEAD,
+    /**
+     * State of adapter instance can't be determined.
+     */
+    UNKNOWN
+}