+++ services/command-router-base/src/main/java/org/eclipse/hono/commandrouter/CommandRouterServiceConfigProperties.java
@@ -18,4 +18,28 @@ import org.eclipse.hono.config.ServiceConfigProperties;
  * Configuration properties for Hono's Command Router service.
  */
 public class CommandRouterServiceConfigProperties extends ServiceConfigProperties {
+
+    private boolean kubernetesBasedAdapterInstanceStatusServiceEnabled = true;
+
+    /**
+     * Checks whether the Kubernetes based service to get the status of an adapter instance is enabled.
+     *
+     * @return {@code true} if the status service is enabled.
+     */
+    public final boolean isKubernetesBasedAdapterInstanceStatusServiceEnabled() {
+        return kubernetesBasedAdapterInstanceStatusServiceEnabled;
+    }
+
+    /**
+     * Sets whether the Kubernetes based service to get the status of an adapter instance should be enabled.
+     *
+     * @param kubernetesBasedAdapterInstanceStatusServiceEnabled {@code true} if the status service should be enabled.
+     * @return This instance for setter chaining.
+     */
+    public final CommandRouterServiceConfigProperties setKubernetesBasedAdapterInstanceStatusServiceEnabled(
+            final boolean kubernetesBasedAdapterInstanceStatusServiceEnabled) {
+        this.kubernetesBasedAdapterInstanceStatusServiceEnabled = kubernetesBasedAdapterInstanceStatusServiceEnabled;
+        return this;
+    }
+
 }